[~] fix: [grains] Grains template now uses u64 type, that can hold (#57)
expected 2^64-1
Why? Right now there is an overflow situation, that accidently doesn't
show up, since vlang treats 2^64 as 0 adding -1 seems to be the same as
2^64-1.

 Zum Commit vorgemerkte Änderungen:
	geändert:       .meta/example.v
 	geändert:       grains.v

+++ exercises/practice/grains/grains.v
@@ -1,9 +1,9 @@
 module main
 
-fn grains_on_square(square i64) !i64 {
+fn grains_on_square(square int) !u64 {
 
 }
 
-fn total_grains_on_board() i64 {
+fn total_grains_on_board() u64 {
 	
 }
